restructured finetuning and pinecone dirs
