+++ align_data/embeddings/pinecone/pinecone_db_handler.py
